Enumerations and Structures
The basic concept of enumerations and structures are easy enough to
understand. Confused on how exactly associated values are assigned
versus how raw values are assigned, further work to be done with
enumerations.

Closure practice tagged on at the end to make sure I still remember it.

Also renamed files for easier access.